Added documentation and moved SafeDeleteQueryset.
