+++ docs/queryset.rst
@@ -1,0 +1,9 @@
+========
+QuerySet
+========
+
+Built-in QuerySet
+-----------------
+
+.. automodule:: safedelete.queryset
+    :members:
